% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{ta_colors}
\alias{ta_colors}
\alias{ta_navy}
\alias{ta_blue}
\alias{ta_lightblue}
\alias{ta_gold}
\alias{ta_muted}
\alias{ta_years_base}
\alias{ta_year_palette}
\title{House-style colours}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 9.
}
\usage{
ta_navy

ta_blue

ta_lightblue

ta_gold

ta_muted

ta_years_base

ta_year_palette(n)
}
\arguments{
\item{n}{Number of colours to return.}
}
\value{
\code{ta_year_palette()} returns a character vector of \code{n} hex colours;
the constants are single hex strings (\code{ta_years_base} is a vector of
nine).
}
\description{
Colour constants used by every plot in the package, plus
\code{ta_year_palette()}, which returns \code{n} distinct colours (interpolated
beyond the nine base colours).
}
\examples{
ta_year_palette(3)
}
\keyword{datasets}
